Better "on weekends" handling (#30)
During review it was noticed that "on weekends" would (at least seemingly)
create an alarm for the next day.  Depending on the phrase, it would
sometimes create a recurring alarm, but nice_date_time() was being used
to describe the alarm and was only formatting the time piece and not the
recurrence.

Also removing unused Padatious entities (old approach) and all Portuguese
vocabulary since it was so out of sync.