18.08 merging VUI changes (#62)
* User experience updates

* Support short alarm dismissal phrases, such as "shut up"
* Prevent duplicate alarms for the same time
* Support "work day" as a recurrance (assumes Mon-Fri)
* Expand vocabulary for "when is my next alarm"
* BUG: Snoozing an alarm could crash with out of range error

* Update test 003

Fix issue with the recurring alarm test

* Workaround skill tester bug

The skill tester chokes on matching dialog that has multiple spaces
in expected dialog.  Removing the double-space.

* Add explicit feedback when deleteing recurring alarms

Added tests and distinct feedback messages when dealing with a recurring
alarm.

* Fix recurring alarms and stopping

* Bug in match_voc() caused blank lines at end of StopBeeping.voc to match
  all phrases, so saying anything stopped the alarm.
* Revamped the alarm dump to include human-readable times
* Fix timezone bug in recurring alarms
* Add "recurring" confirmation when Delete All only finds a single recurring alarm

* Fix dialg in test

* Support non-supported "except"

Inform the user that you can't say "except" in a recurring alarm.  We
will likely handle this in the future.

* Add support for multiple days in repeats

Now the user can speak combinations of day groups, e.g.
"set an alarm at 7pm on Monday and Wednesday" or
"set an alarm at 8am on weekends and Wednesdays"

* Fix initial recurrance report

Setting a recurring alarm for the same day was creating a weird
confirmation description.  E.g. "set an alarm for 7am on Fridays"
when run on a Friday at noon would confirm with "I've created
an alarm for 7am on Fridays, starting now from now"

* Document a difficult edge case for future debugging

* Add support for "private.mycroftai.has_alarm" messagebus message

This can be used by the digital clock face as an indicator