Bugfix: initial recurring alarm
A future recurring alarm would have it's first instance set based on the requested _time_,
not on the time and the future dates.  For example, on a Wednesday afternoon if you said:
```set and alarm for 8am on weekends```
It would calculate "8am" as "the next 8am after right now", meaning 8am on Thursday.  So
the first instance of the alarm would trigger on Thursday morning, not Saturday.  After
the first instance it was fine.

Fixed by doing a little calendar math and then running the recurrance rule.

Also:
* Switched to using 19.02's mycroft.util.format.join()
* Prepared nice_relative_time() to be moved into mycroft-core
* Add intent vocabulary to catch "when is the next alarm set?" and such