Remove disable_listen_beep
This was intended to make it quicker to stop an alarm, by not
playing the listening beep. However this being inconsistent causes
UX issues where users wait for the beep, the listener times out
and then doesn't hear their eventual stop. It's also bad practice
to be manually patching the config continually.