Update dismissal validator when asking to set alarm without time.
Primary issue was that the get_response method was using
extract_datetime as a validator. Therefore only responses that
contained an extractable datetime could continue code execution.
This meant the existing check of the dismiss vocab would never
find anything of significance. Only the vocab included in mycroft
-core was caught.

2nd, the test expected 'no' to dismiss an interaction. However it
doesn't make sense in the context of the question, and has a high
chance of causing issues, if STT returned 'no' as part of an alarm
name or something else.

A range of other vocab options have been added.