Fix tests failing on specific days
Because alarms were being set on specific days of the current
week there is the potential that the test runs at a time where
the behaviour should differ. Eg the alarm may be set then expire

This changes test alarms to be either tomorrow or a day next
week. Therefore it cannot be today.